Added Calendar Application Progress Button saving function under program academics view plus Bug fixes. Reached 40,000 lines milestone

+++ JumpPad/iOSDateView.m
@@ -74,7 +74,7 @@
     
     
     //Month
-    NSString* monthString = [[JPGlobal monthStringWithInt:self.month] uppercaseString];
+    NSString* monthString = [[JPGlobal monthStringWithInt:(int)self.month] uppercaseString];
     
     CGRect monthRect = CGRectMake(contentRect.origin.x, rect.size.height* 0.13, contentRect.size.width, rect.size.height* 0.2);
